Resize of time to allow for more time expression

+++ wrappers.c
@@ -127,7 +127,7 @@ void dice(struct discord *client, const struct discord_message *event, char *ms,
 }
 
 time_t sec_convert(char *s){ // converts time (<y, M, w, d, h, m, s> + int) to seconds
-	char unit; int value;
+	char unit; long long unsigned int value;
 	sscanf(s, "%c%d", &unit, &value);
 	switch(unit){
 		case 'y':
